Хранение вьюмоделей основных экранов
