+++ lib/core/viewmodels/factories/main_page_routes_view_models_factory.dart
@@ -1,0 +1,26 @@
+import 'package:injector/injector.dart';
+import 'package:unn_mobile/core/services/interfaces/authorisation_service.dart';
+import 'package:unn_mobile/core/viewmodels/base_view_model.dart';
+
+class MainPageRoutesViewModelsFactory {
+  final AuthorizationService _authorizationService;
+
+  final Map<int, BaseViewModel> _viewModels = {};
+
+  MainPageRoutesViewModelsFactory(this._authorizationService) {
+    _authorizationService.addListener(
+      _resetIfUnauthorized,
+    ); // А где отписаться? Я хз) Это должен быть синглтон, надеюсь, ничего не сломается
+  }
+  T getViewModelByRouteIndex<T extends BaseViewModel>(int index) {
+    _viewModels.putIfAbsent(index, () => Injector.appInstance.get<T>());
+    return _viewModels[index] as T;
+  }
+
+  void _resetIfUnauthorized() {
+    if (_authorizationService.isAuthorised) {
+      return;
+    }
+    _viewModels.clear();
+  }
+}
