+++ lib/core/viewmodels/main_page_route_view_model.dart
@@ -1,0 +1,3 @@
+abstract interface class MainPageRouteViewModel {
+  void refresh();
+}
