+++ lib/core/viewmodels/main_page_view_model.dart
@@ -1,4 +1,7 @@
+import 'package:injector/injector.dart';
 import 'package:unn_mobile/core/viewmodels/base_view_model.dart';
+import 'package:unn_mobile/core/viewmodels/factories/main_page_routes_view_models_factory.dart';
+import 'package:unn_mobile/core/viewmodels/main_page_route_view_model.dart';
 import 'package:unn_mobile/core/viewmodels/profile_view_model.dart';
 
 class MainPageViewModel extends BaseViewModel {
@@ -9,7 +12,19 @@ class MainPageViewModel extends BaseViewModel {
   ProfileViewModel get profileViewModel => _profileViewModel;
 
   void init() {
-    setState(ViewState.busy);
+    if (isInitialized) {
+      return;
+    }
+    isInitialized = true;
     _profileViewModel = ProfileViewModel.currentUser();
+  }
+
+  void refreshTab(int index) {
+    final tab = Injector.appInstance
+        .get<MainPageRoutesViewModelsFactory>()
+        .getViewModelByRouteIndex(index);
+    if (tab is MainPageRouteViewModel) {
+      (tab as MainPageRouteViewModel).refresh();
+    }
   }
 }
